backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "black>=25.9.0",
    "coloredlogs>=15.0.1",
    "fastapi[standard]>=0.117.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "ruff>=0.13.2",
]
